# backend/openapi-schema.yml
components:
  schemas:
...
    ChangePasswordRequest:
      properties:
        new_password:
          minLength: 6
          title: New Password
          type: string
        old_password:
          title: Old Password
          type: string
      required:
      - old_password
      - new_password
      title: ChangePasswordRequest
      type: object
...
    CreateUser:
      properties:
        password:
          title: Password
          type: string
        username:
          title: Username
          type: string
      required:
      - username
      - password
      title: CreateUser
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    LoginResponse:
      properties:
        token:
          title: Token
          type: string
      required:
      - token
      title: LoginResponse
      type: object
...
    UserBase:
      properties:
        username:
          title: Username
          type: string
      required:
      - username
      title: UserBase
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    HTTPBasic:
      scheme: basic
      type: http
info:
  title: FastAPI
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/users/change_password/:
    post:
      operationId: change_password_api_v1_users_change_password__post
      parameters:
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBase'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Change Password
  /api/v1/users/create/:
    post:
      operationId: create_user_req_api_v1_users_create__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create User Req
  /api/v1/users/login/:
    post:
      operationId: login_api_v1_users_login__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Login
  /api/v1/users/logout/:
    post:
      operationId: logout_api_v1_users_logout__post
      parameters:
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Logout
  /api/v1/users/me/:
    get:
      operationId: read_user_api_v1_users_me__get
      parameters:
      - in: header
        name: authorization
        required: true
        schema:
          title: Authorization
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBase'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Read User
...